Recreated 3 main pages in React
Recreated our page's skeletons for Main, Friends, and Profile Pages with CSS styling and a navigation bar in React. The main page contains a new entry button which the other pages do not have. We utilize React Router and Styled-Components for styling and having multiple pages. The Component Folder contains the Navigation bar used in all pages. The Pages Folder contains the individual pages.

Co-Authored-By: liz-areizaga <81609405+liz-areizaga@users.noreply.github.com>